refactor out main menu
